Move passphrase hooks into own file
