+++ wallets/client/hooks/index.js
@@ -7,3 +7,4 @@ export * from './crypto'
 export * from './query'
 export * from './logger'
 export * from './diagnostics'
+export * from './passphrase'
